requirements.txt
argparse
httplib2
pycrypto
requests
urllib3
